requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
qrcode==8.2
Werkzeug==3.1.3
qrcode[pil]
ddtrace==2.10.0
flask_sqlalchemy
psycopg2-binary
requests